refactor: move idl ts objects inside releases
