+++ src/solana/releases/airdrops/v0.1/aliases.ts
@@ -1,0 +1,5 @@
+import manifest from "./manifest";
+
+const aliases = { [manifest.SABLIER_MERKLE_INSTANT]: "merkleFactoryInstant" };
+
+export default aliases;
