+++ src/solana/releases/airdrops/v0.1/deployments.ts
@@ -11,7 +11,7 @@ function get(chainId: number, programMap: Sablier.Solana.ProgramMap): Sablier.So
     chainId,
     contractMap: programMap,
     protocol: Protocol.Airdrops,
-    version: "v1.0",
+    version: "v0.1",
   });
 }
