+++ src/solana/releases/airdrops/v0.1/idl/SablierMerkleInstant/structs.ts
