+++ src/solana/releases/airdrops/v0.1/manifest.ts
@@ -1,0 +1,3 @@
+const manifest = { SABLIER_MERKLE_INSTANT: "SablierMerkleInstant" };
+
+export default manifest;
