--- src/solana/releases/airdrops/v1.0/aliases.ts
@@ -1,7 +1,0 @@
-import manifest from "./manifest";
-
-const aliases = {
-  [manifest.SABLIER_MERKLE_INSTANT]: "merkleFactoryInstant",
-};
-
-export default aliases;
