--- src/solana/releases/airdrops/v1.0/manifest.ts
@@ -1,5 +1,0 @@
-const manifest = {
-  SABLIER_MERKLE_INSTANT: "SablierMerkleInstant",
-};
-
-export default manifest;
