+++ src/solana/releases/biome.jsonc
@@ -1,0 +1,21 @@
+{
+  "$schema": "../../../node_modules/@biomejs/biome/configuration_schema.json",
+  "extends": "//",
+  "overrides": [
+    {
+      "includes": ["**/*.{json,jsonc,ts}"],
+      "assist": {
+        "actions": {
+          "source": {
+            "useSortedKeys": "off"
+          }
+        }
+      },
+      "javascript": {
+        "formatter": {
+          "expand": "never" // Never expand objects in the IDL to save number of lines
+        }
+      }
+    }
+  ]
+}
