+++ src/solana/releases/data.ts
@@ -1,7 +1,4 @@
 import { airdrops } from "./airdrops";
 import { lockup } from "./lockup";
 
-export const releases = {
-  airdrops,
-  lockup,
-};
+export const releases = { airdrops, lockup };
