+++ src/solana/releases/helpers.ts
@@ -1,5 +1,4 @@
 import { chainsQueries } from "@src/solana/chains/queries";
-import _ from "lodash";
 
 export function sortDeployments<T extends { chainId: number }>(deployments: T[]): T[] {
   return deployments.sort((a, b) => {
