+++ src/solana/releases/lockup/index.ts
@@ -1,6 +1,4 @@
 import type { Sablier } from "@src/types";
-import { release as releaseV1_0 } from "./v1.0";
+import { release as releaseV0_1 } from "./v0.1";
 
-export const lockup: Record<Sablier.Solana.Version.Lockup, Sablier.Solana.Release> = {
-  "v1.0": releaseV1_0,
-};
+export const lockup: Record<Sablier.Solana.Version.Lockup, Sablier.Solana.Release> = { "v0.1": releaseV0_1 };
