+++ src/solana/releases/lockup/v0.1/idl/SablierLockupLinear/errors.ts
