+++ src/solana/releases/lockup/v0.1/manifest.ts
@@ -1,0 +1,3 @@
+const manifest = { SABLIER_LOCKUP_LINEAR: "SablierLockupLinear" };
+
+export default manifest;
