--- src/solana/releases/lockup/v1.0/aliases.ts
@@ -1,7 +1,0 @@
-import manifest from "./manifest";
-
-const aliases = {
-  [manifest.SABLIER_LOCKUP_LINEAR]: "LL",
-};
-
-export default aliases;
