--- src/solana/releases/lockup/v1.0/manifest.ts
@@ -1,5 +1,0 @@
-const manifest = {
-  SABLIER_LOCKUP_LINEAR: "SablierLockupLinear",
-};
-
-export default manifest;
